Plotting in separate class
